Add delete method for backupbucket
